Fix ppp kernel build. Closes #1
In the previous commit, I tried staying on the rc3 kernel,
because there use dto be problems with the display with the newest one at the time.
I used the commit hash for that. I still had the commit in the local copy
of the repo, but it seams it wasn't present for people who newly cloned
it anymore.